+++ 2DAction/Source/Game/Enemy/EnemySlime.h
@@ -23,8 +23,8 @@ protected:
 	
 	virtual bool InitMain() override;									// 派生先での初期化
 	virtual const Common::TYPE_OBJECT GetTypeObject() const override{ return Common::TYPE_ENEMY_SLIME;}
-	virtual const uint32_t GetEnemyDefaultHP() const override;			// HP取得
-	virtual const uint32_t GetEnemyDefaultSPD() const override;			// SPD取得
+	virtual const uint32_t	GetEnemyDefaultHP() const override;			// HP取得
+	virtual const float		GetEnemyDefaultSPD() const override;			// SPD取得
 	virtual void  ReduceDamage( Common::CMN_EVENT &eventId ) override;	// 敵の種類、レベル等でダメージ軽減処理
 
 private:
